pyproject.toml
[project]
name = "probase2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=23.0.0",
    "aiosqlite>=0.19.0",
    "bcrypt==3.2.2",
    "fastapi[all]>=0.104.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.25.0",
    "itsdangerous>=2.1.2",
    "jinja2>=3.1.2",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.9",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "mem0ai>=1.0.4",
    "neo4j>=6.1.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "pypdf>=6.7.0",
    "pyproj>=3.7.2",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27 ; sys_platform != 'win32'",
    "python-magic-bin>=0.4.14",
    "python-magic-bin>=0.4.14 ; sys_platform == 'win32'",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.16.2",
    "redis>=7.1.1",
    "rq>=2.6.1",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]
